Refactor backend: Modularize main.py, add config and routers
