+++ backend/routers/__init__.py
@@ -1,0 +1,1 @@
+# Routers package for organizing FastAPI endpoints by feature area
