+++ backend/routers/core.py
@@ -1,0 +1,128 @@
+from fastapi import APIRouter, HTTPException, Depends
+from typing import List
+import httpx
+
+from ..models import Agent, System, Waypoint
+from ..config import HAS_VALID_TOKEN, SPACETRADERS_API_URL, SPACETRADERS_TOKEN
+from ..mock_data import MOCK_AGENT, MOCK_SYSTEMS, MOCK_WAYPOINTS, MOCK_FACTIONS
+from ..utilities import get_httpx_client
+
+router = APIRouter(prefix="/api", tags=["core"])
+
+@router.get("/status")
+async def get_status():
+    """Get SpaceTraders API status"""
+    try:
+        async with httpx.AsyncClient() as client:
+            response = await client.get(f"{SPACETRADERS_API_URL}")
+            return response.json()
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+@router.get("/agent", response_model=Agent)
+async def get_agent(client: httpx.AsyncClient = Depends(get_httpx_client)):
+    """Get current agent information"""
+    if not HAS_VALID_TOKEN:
+        return MOCK_AGENT
+    
+    try:
+        headers = {"Authorization": f"Bearer {SPACETRADERS_TOKEN}"}
+        response = await client.get(f"{SPACETRADERS_API_URL}/my/agent", headers=headers)
+        
+        if response.status_code == 200:
+            data = response.json()
+            return data["data"]
+        else:
+            raise HTTPException(status_code=response.status_code, detail=response.text)
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+@router.get("/systems", response_model=List[System])
+async def get_systems(client: httpx.AsyncClient = Depends(get_httpx_client)):
+    """Get all systems"""
+    if not HAS_VALID_TOKEN:
+        return MOCK_SYSTEMS
+    
+    try:
+        headers = {"Authorization": f"Bearer {SPACETRADERS_TOKEN}"}
+        response = await client.get(f"{SPACETRADERS_API_URL}/systems", headers=headers)
+        
+        if response.status_code == 200:
+            data = response.json()
+            return data["data"]
+        else:
+            raise HTTPException(status_code=response.status_code, detail=response.text)
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+@router.get("/factions")
+async def get_factions(client: httpx.AsyncClient = Depends(get_httpx_client)):
+    """Get all factions"""
+    if not HAS_VALID_TOKEN:
+        return MOCK_FACTIONS
+    
+    try:
+        headers = {"Authorization": f"Bearer {SPACETRADERS_TOKEN}"}
+        response = await client.get(f"{SPACETRADERS_API_URL}/factions", headers=headers)
+        
+        if response.status_code == 200:
+            data = response.json()
+            return data["data"]
+        else:
+            raise HTTPException(status_code=response.status_code, detail=response.text)
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+@router.get("/info")
+async def get_info():
+    """Get application info and token status"""
+    return {
+        "has_valid_token": HAS_VALID_TOKEN,
+        "token_configured": bool(SPACETRADERS_TOKEN),
+        "api_url": SPACETRADERS_API_URL,
+        "message": "Use demo data" if not HAS_VALID_TOKEN else "Using real SpaceTraders API"
+    }
+
+@router.get("/systems/{system_symbol}/waypoints", response_model=List[Waypoint])
+async def get_system_waypoints(system_symbol: str, client: httpx.AsyncClient = Depends(get_httpx_client)):
+    """Get all waypoints in a system"""
+    if not HAS_VALID_TOKEN:
+        # Return mock waypoints for the demo system
+        if system_symbol == "X1-DF55":
+            return MOCK_WAYPOINTS
+        else:
+            return []
+    
+    try:
+        headers = {"Authorization": f"Bearer {SPACETRADERS_TOKEN}"}
+        response = await client.get(f"{SPACETRADERS_API_URL}/systems/{system_symbol}/waypoints", headers=headers)
+        
+        if response.status_code == 200:
+            data = response.json()
+            return data["data"]
+        else:
+            raise HTTPException(status_code=response.status_code, detail=response.text)
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+@router.get("/systems/{system_symbol}", response_model=System)
+async def get_system(system_symbol: str, client: httpx.AsyncClient = Depends(get_httpx_client)):
+    """Get system details"""
+    if not HAS_VALID_TOKEN:
+        # Return mock system for demo
+        if system_symbol == "X1-DF55":
+            return MOCK_SYSTEMS[0]
+        else:
+            raise HTTPException(status_code=404, detail="System not found")
+    
+    try:
+        headers = {"Authorization": f"Bearer {SPACETRADERS_TOKEN}"}
+        response = await client.get(f"{SPACETRADERS_API_URL}/systems/{system_symbol}", headers=headers)
+        
+        if response.status_code == 200:
+            data = response.json()
+            return data["data"]
+        else:
+            raise HTTPException(status_code=response.status_code, detail=response.text)
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
