+++ backend/routers/scanning.py
@@ -1,0 +1,155 @@
+from fastapi import APIRouter, HTTPException, Depends
+import httpx
+
+from ..config import HAS_VALID_TOKEN, SPACETRADERS_API_URL, SPACETRADERS_TOKEN
+from ..mock_data import MOCK_SCAN_RESULTS, MOCK_SURVEYS
+from ..utilities import get_httpx_client
+
+router = APIRouter(prefix="/api/ships", tags=["scanning"])
+
+@router.post("/{ship_symbol}/scan/systems")
+async def scan_systems(ship_symbol: str, client: httpx.AsyncClient = Depends(get_httpx_client)):
+    """Long-range sensors - Detect systems and celestial objects"""
+    if not HAS_VALID_TOKEN:
+        # Mock response for demo
+        return {
+            "data": {
+                "cooldown": {
+                    "shipSymbol": ship_symbol,
+                    "totalSeconds": 70,
+                    "remainingSeconds": 70,
+                    "expiration": "2023-11-01T00:01:10.000Z"
+                },
+                "systems": MOCK_SCAN_RESULTS["systems"]
+            }
+        }
+    
+    try:
+        headers = {"Authorization": f"Bearer {SPACETRADERS_TOKEN}"}
+        response = await client.post(f"{SPACETRADERS_API_URL}/my/ships/{ship_symbol}/scan/systems", headers=headers)
+        
+        if response.status_code == 201:
+            return response.json()
+        else:
+            raise HTTPException(status_code=response.status_code, detail=response.text)
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+@router.post("/{ship_symbol}/scan/waypoints")
+async def scan_waypoints(ship_symbol: str, client: httpx.AsyncClient = Depends(get_httpx_client)):
+    """Planetary survey - Scan waypoints for resources and composition"""
+    if not HAS_VALID_TOKEN:
+        # Mock response for demo
+        return {
+            "data": {
+                "cooldown": {
+                    "shipSymbol": ship_symbol,
+                    "totalSeconds": 60,
+                    "remainingSeconds": 60,
+                    "expiration": "2023-11-01T00:01:00.000Z"
+                },
+                "waypoints": MOCK_SCAN_RESULTS["waypoints"]
+            }
+        }
+    
+    try:
+        headers = {"Authorization": f"Bearer {SPACETRADERS_TOKEN}"}
+        response = await client.post(f"{SPACETRADERS_API_URL}/my/ships/{ship_symbol}/scan/waypoints", headers=headers)
+        
+        if response.status_code == 201:
+            return response.json()
+        else:
+            raise HTTPException(status_code=response.status_code, detail=response.text)
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+@router.post("/{ship_symbol}/scan/ships")
+async def scan_ships(ship_symbol: str, client: httpx.AsyncClient = Depends(get_httpx_client)):
+    """Signal interception and threat assessment - Scan nearby ships"""
+    if not HAS_VALID_TOKEN:
+        # Mock response for demo
+        return {
+            "data": {
+                "cooldown": {
+                    "shipSymbol": ship_symbol,
+                    "totalSeconds": 10,
+                    "remainingSeconds": 10,
+                    "expiration": "2023-11-01T00:00:10.000Z"
+                },
+                "ships": MOCK_SCAN_RESULTS["ships"]
+            }
+        }
+    
+    try:
+        headers = {"Authorization": f"Bearer {SPACETRADERS_TOKEN}"}
+        response = await client.post(f"{SPACETRADERS_API_URL}/my/ships/{ship_symbol}/scan/ships", headers=headers)
+        
+        if response.status_code == 201:
+            return response.json()
+        else:
+            raise HTTPException(status_code=response.status_code, detail=response.text)
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+@router.post("/{ship_symbol}/survey")
+async def create_survey(ship_symbol: str, client: httpx.AsyncClient = Depends(get_httpx_client)):
+    """Resource mapping - Create detailed survey of current waypoint"""
+    if not HAS_VALID_TOKEN:
+        # Mock response for demo
+        return {
+            "data": {
+                "cooldown": {
+                    "shipSymbol": ship_symbol,
+                    "totalSeconds": 60,
+                    "remainingSeconds": 60,
+                    "expiration": "2023-11-01T00:01:00.000Z"
+                },
+                "surveys": MOCK_SURVEYS
+            }
+        }
+    
+    try:
+        headers = {"Authorization": f"Bearer {SPACETRADERS_TOKEN}"}
+        response = await client.post(f"{SPACETRADERS_API_URL}/my/ships/{ship_symbol}/survey", headers=headers)
+        
+        if response.status_code == 201:
+            return response.json()
+        else:
+            raise HTTPException(status_code=response.status_code, detail=response.text)
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
+
+@router.get("/{ship_symbol}/cooldown")
+async def get_ship_cooldown(ship_symbol: str, client: httpx.AsyncClient = Depends(get_httpx_client)):
+    """Get current ship cooldown status"""
+    if not HAS_VALID_TOKEN:
+        # Mock response - no cooldown
+        return {
+            "data": {
+                "shipSymbol": ship_symbol,
+                "totalSeconds": 0,
+                "remainingSeconds": 0,
+                "expiration": None
+            }
+        }
+    
+    try:
+        headers = {"Authorization": f"Bearer {SPACETRADERS_TOKEN}"}
+        response = await client.get(f"{SPACETRADERS_API_URL}/my/ships/{ship_symbol}/cooldown", headers=headers)
+        
+        if response.status_code == 200:
+            return response.json()
+        elif response.status_code == 404:
+            # No cooldown
+            return {
+                "data": {
+                    "shipSymbol": ship_symbol,
+                    "totalSeconds": 0,
+                    "remainingSeconds": 0,
+                    "expiration": None
+                }
+            }
+        else:
+            raise HTTPException(status_code=response.status_code, detail=response.text)
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=str(e))
